requirements.txt
anyio==3.3.4
asgiref==3.4.1
boto3==1.22.13
botocore==1.25.13
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Cython==0.29.28
fastapi==0.70.0
Flask==2.0.2
h11==0.12.0
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==1.0.0
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.4.58
Pillow==8.4.0
pydantic==1.8.2
pytesseract==0.3.8
python-dateutil==2.8.2
python-multipart==0.0.5
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.9
uvicorn==0.15.0
Werkzeug==2.0.2
wincertstore==0.2
